[*]拆分加载ab部分和普通加载.
