+++ Assets/Main/LoadDllFromAssetbundle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49555ce87d310f747b31384b8ef95af5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
